uHAL : Fix code leading to compiler warnings (integer-unsigned comparisons) with gcc 7.2.1; refs #65

+++ uhal/tests/src/common/test_hierarchy.cpp
@@ -45,7 +45,7 @@
 #include <typeinfo>
 
 
-#define N_1MB    1024*1024/4
+#define N_1MB    uint32_t(1024*1024/4)
 
 
 namespace uhal {
